Indexed should be ints, not longs
+ The odds that somebody will have a 2^32 length stream are remote, and the most common case should be the easiest to implement

+++ src/main/java/com/ginsberg/gatherers4j/IndexedAccumulatorFunction.java
@@ -30,5 +30,5 @@ public interface IndexedAccumulatorFunction<
     /// @param carry the accumulated value
     /// @param next  the next value to accumulate
     /// @return the function result
-    R apply(long index, @Nullable A carry, @Nullable B next);
+    R apply(int index, @Nullable A carry, @Nullable B next);
 }
